refactor: Moved dependency functions to _crud.py & _deps.py
